BD nueva + otros menesteres

+++ src/controlador/Especialidad/ControladorVtnElimEspecialidad.java
@@ -2,9 +2,15 @@ package controlador.Especialidad;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.sql.SQLException;
+import java.util.ArrayList;
+import java.util.List;
+
+import javax.swing.JOptionPane;
 
 import modelo.Especialidad.Especialidad;
 import modelo.Especialidad.EspecialidadBD;
+import modelo.Medico.Medico;
+import modelo.Medico.MedicoBD;
 import modelo.Paciente.PacienteBD;
 import vista.Especialidad.VentanaEliminarEspecialidad;
 
@@ -43,10 +49,20 @@ public class ControladorVtnElimEspecialidad implements ActionListener {
 	private void eliminarEspecialidad(){
 	  	try
 		{
-	  		EspecialidadBD especialidadBD = new EspecialidadBD();	    
-	    	especialidadBD.eliminarEspecialidad(vtnElimEsp.getCodigo());
-	    	vtnElimEsp.mostrarMensaje("La especialidad fue eliminado con exito");
-	    	vtnElimEsp.blanquearCampos();
+			Object[] opciones = {"Sí","No"};
+	  		int n = JOptionPane.showOptionDialog(null,"Eliminar una especialidad eliminará a todos los médicos asociados a la misma. \n¿Desea continuar?","Eliminar especialidad",JOptionPane.YES_NO_OPTION,JOptionPane.QUESTION_MESSAGE,null,opciones,opciones[1]);
+			if (n==JOptionPane.YES_OPTION) {
+		  		EspecialidadBD especialidadBD = new EspecialidadBD();	  
+		  		List<Medico> medicos = new ArrayList<Medico>();
+		  		MedicoBD medicoBD = new MedicoBD();
+		    	especialidadBD.eliminarEspecialidad(vtnElimEsp.getCodigo());
+		    	medicos = medicoBD.consultarFiltrarMedicos("cod_especialidad = '"+vtnElimEsp.getCodigo()+"'");
+		    	for (int i=0;i<medicos.size();i++) {
+		    		medicoBD.eliminarMedico(medicos.get(i).getCedula());
+		    	}
+		    	vtnElimEsp.mostrarMensaje("La especialidad fue eliminado con exito");
+		    	vtnElimEsp.blanquearCampos();
+			}
 		}catch(Exception e)
 		{
 			vtnElimEsp.mostrarMensaje("No se pudo eliminar el Paciente, verifique que los datos sean correctos");
